Removed duplicate functions.  Added error checking.  Fixed one bug that would cause crash on launch.
git-svn-id: svn://localhost/media/c/Documnts/src/flare/repo.svndump/trunk@149 6ae5efe3-03d3-4ed7-8a64-3d71375a9a67

+++ src/UtilsParsing.h
@@ -18,7 +18,7 @@ string trim(string s, char c);
 string parse_section_title(string s);
 void parse_key_pair(string s, string &key, string &val);
 int eatFirstInt(string &s, char separator);
-unsigned eatFirstHex(string &s, char separator);
+unsigned short eatFirstHex(string &s, char separator);
 string eatFirstString(string &s, char separator);
 
 #endif
